I added tags
i added tags to the things that were "untagged" so that i can use more specific yet simpler code in the future if i want to fx. the fences if i want the Cow(player) to losses damage (fx. 'cause of electricity if i change the fences to electric fences on this level or another) every time they touche the fence the fences tags are Fences.