Added vampire spawning system(NotWoking... yet)
I've tried to add a vampire spawning system, but it doesn't work as expected. Whenever I run the program, 45 seconds go by (as expected), and then instead of spawning 1 vampire, it spawns 10! Afterwards 1 spawns every 5 seconds. I currently have no idea, how to fix this issue, but I'm up for the challenge, and I do think that I can solve it in the future.